refactor(auth): unify apps/api on the worker better-auth copy, collapse the oauth-seed workaround
The duplicate-package landmine was misdiagnosed as svelte-vs-pg peer sets. The
only optional peer of @better-auth/core is @cloudflare/workers-types, so that
is the sole axis that forks a second copy. packages/server and apps/team-api
both declare it and resolve the same copy (+6e2c8aa4); apps/api alone omitted
it (it uses `wrangler types`) and resolved +4b1fb26a, so its program mixed two
@better-auth/core copies and produced the AuthContext cast error whenever it
pulled @epicenter/server's auth types. apps/team-api is the control group: same
server auth surface, never hit the whack-a-mole, because it declares the dep.

Declare @cloudflare/workers-types in apps/api, matching server and team-api
verbatim. apps/api now canonicalizes to +6e2c8aa4, so every worker deployment
program agrees on one copy. That removes the reason the seed workaround existed:

  - Move projectTrustedOAuthClientToRow into @epicenter/constants/oauth, beside
    buildTrustedOAuthClients (its input). The seed script and the auth test now
    reach it without importing the request-path barrel.
  - Delete the bespoke @epicenter/server/oauth-clients entry and its source
    file; the seed no longer depends on @epicenter/server at all.
  - Rename trusted-oauth-clients.test.ts to plugins.test.ts, since it exercises
    authPlugins now that its namesake source is gone.

The seed keeps its raw parameterized upsert: drizzle-orm is still duplicated
(a separate axis the workers-types fix does not touch), so a drizzle query
across the apps/api/server boundary would still mismatch copies.

Both worker typechecks (@epicenter/server, apps/api) and @epicenter/constants
pass; the auth tests are green.
