+++ apps/api/package.json
@@ -17,6 +17,8 @@
 		"typegen": "wrangler types",
 		"auth:generate": "infisical run --env=prod --path=/ops --path=/api -- bun x @better-auth/cli generate --yes --config ./better-auth.config.ts --output ../../packages/server/src/db/schema/auth.ts",
 		"db:generate": "drizzle-kit generate",
+		"oauth:seed:local": "bun scripts/seed-oauth-clients.ts",
+		"oauth:seed:remote": "infisical run --env=prod --path=/ops -- env SEED_TARGET=prod bun scripts/seed-oauth-clients.ts",
 		"db:drop": "drizzle-kit drop",
 		"db:push:local": "drizzle-kit push",
 		"db:migrate:remote": "infisical run --env=prod --path=/ops -- drizzle-kit migrate",
@@ -50,6 +52,7 @@
 		"yjs": "catalog:"
 	},
 	"devDependencies": {
+		"@cloudflare/workers-types": "^4.20260405.1",
 		"@types/bun": "catalog:",
 		"@types/pg": "^8.18.0",
 		"atmn": "^1.1.5",
